Add server side route redirection
Add middleware function which checks if the JWT is saved in the HTTP Only cookie.
The isAuth function in the client side handles the get fetch request to the server on page load.
If the token doesnt exist, send a redirect response to the login page, else save the user into the reqest.